move config file to CLI folder under .huck
huck create <CLI> now creates a folder under .huck instead of a
freestanding configuration file directly under .huck. This allows
for configuration groupings for individual CLIs.

Creation of a configuration folder for a CLI also now creates a config
file, sets a default section and and a url to empty.