cleanup README and cleaner .bash_profile and huck_profile behavior
Additional checks are put in place to ensure that the .bash_profile and
the huck_profile don't get repeating entries if huck create is used
multiple time for the same cli name.

We now also check to see if a configuation file for a given cli already
exists and we leave it alone if it does.